show current weather with location permission imp.

+++ app/src/main/kotlin/com/varol/weathever/data/remote/model/BaseResponseModel.kt
@@ -1,3 +1,3 @@
 package com.varol.weathever.data.remote.model
 
-open class BaseResponseModel(val cod: Int)
+open class BaseResponseModel(val cod: Long)
